# docs/swagger.yaml
definitions:
  http.postItem:
    properties:
      content:
        type: string
      file_ids:
        items:
          type: string
        type: array
      id:
        type: string
      pin:
        type: boolean
      tagged_target:
        items:
          type: string
        type: array
    type: object
  http.usersItem:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_post_delivery_http.createReq:
    properties:
      content:
        type: string
      file_ids:
        items:
          type: string
        type: array
      pin:
        type: boolean
      tagged_target:
        items:
          type: string
        type: array
    type: object
  internal_post_delivery_http.detailResp:
    properties:
      content:
        type: string
      file_ids:
        items:
          type: string
        type: array
      id:
        type: string
      pin:
        type: boolean
      tagged_target:
        items:
          type: string
        type: array
    type: object
  internal_post_delivery_http.getMetaResponse:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_post_delivery_http.getResp:
    properties:
      items:
        items:
          $ref: '#/definitions/http.postItem'
        type: array
      meta:
        $ref: '#/definitions/internal_post_delivery_http.getMetaResponse'
    type: object
  internal_post_delivery_http.updateReq:
    properties:
      content:
        type: string
      file_ids:
        items:
          type: string
        type: array
      id:
        type: string
      tagged_target:
        items:
          type: string
        type: array
    type: object
  internal_users_delivery_http.createReq:
    properties:
      birthday:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  internal_users_delivery_http.detailResp:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_users_delivery_http.getMetaResponse:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_users_delivery_http.getResp:
    properties:
      items:
        items:
          $ref: '#/definitions/http.usersItem'
        type: array
      meta:
        $ref: '#/definitions/internal_users_delivery_http.getMetaResponse'
    type: object
  internal_users_delivery_http.updateReq:
    properties:
      avatar_url:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  response.Resp:
    properties:
      data: {}
      error_code:
        type: integer
      errors: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/news-feed/posts:
    get:
      consumes:
      - application/json
      description: Get post with filter
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJvIjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: query
        name: id
        type: string
      - collectionFormat: csv
        description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      - description: Type
        enum:
        - company
        - self
        in: query
        name: type
        type: string
      - description: Status
        enum:
        - approved
        - pending
        - rejected
        - pending_update
        in: query
        name: status
        type: string
      - description: Pin
        in: query
        name: pin
        type: boolean
      - description: User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_post_delivery_http.getResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get post
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update post
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_post_delivery_http.updateReq'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_post_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Update post
      tags:
      - Posts
  /api/v1/news-feed/posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete post
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_post_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Delete post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Get post detail
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_post_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get post detail
      tags:
      - Posts
  /news-feed/posts:
    post:
      consumes:
      - application/json
      description: Create post
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_post_delivery_http.createReq'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_post_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Create post
      tags:
      - Posts
  /news-feed/user:
    get:
      consumes:
      - application/json
      description: Get user with filter
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJvIjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: query
        name: id
        type: string
      - collectionFormat: csv
        description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      - description: Type
        enum:
        - company
        - self
        in: query
        name: type
        type: string
      - description: Status
        enum:
        - approved
        - pending
        - rejected
        - pending_update
        in: query
        name: status
        type: string
      - description: Pin
        in: query
        name: pin
        type: boolean
      - description: User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users_delivery_http.getResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get user
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_users_delivery_http.createReq'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Create user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_users_delivery_http.updateReq'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Update user
      tags:
      - Users
  /news-feed/user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user detail
      parameters:
      - default: '*'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Swagger-Codegen/1.0.0/go
        description: User-Agent
        in: header
        name: User-Agent
        type: string
      - default: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vcC50YW5jYS52bi9hcGkvdjQvYXV0aC9zaWduaW4tdjIiLCJpYXQiOjE3MTY1ODUyNDAsIm5iZiI6MTcxNjU4NTI0MCwianRpIjoidFBJMldUa0JldThYdnJMZiIsInN1YiI6Ik5pdEpwZUp1dkF4M1pjYXdKIiwicHJ2IjoiMWM1NTIwZjcwYmFhNjU1ZGRjNTc0NmE2NzY0ZjM3MmExYjY1NWFhNiIsInNob3BfaWQiOiI1YzIwYTE5YzBiMDg4ODBmNTk0ZmM0NjgiLCJzaG9wX3VzZXJuYW1lIjoicmF2ZSIsInNob3BfcHJlZml4IjoidCIsInR5cGUiOiJhcGkifQ.DnxirM5IXQY3B9Vcc6Qco7c9f0ABGjoeLu_1LfHiRjE
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: en
        description: Language
        in: header
        name: lang
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_users_delivery_http.detailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get user detail
      tags:
      - Users
swagger: "2.0"